ismn5650-lrd0036/Assignment4/requirements.txt
pandas==2.1.0
python-dotenv==1.0.0
requests==2.31.0
